[apache] If SSL enabled, enforce redirection from HTTP to HTTPS.
It is assumed that if a user starts the container with the SSL option,
he wants to disable the access through HTTP and prefers to ensure a
secure connection. Therefore, if SSL is enabled, redirection to the
https:// url is enforced.

.htaccess is used instead of pure VirtualHost redirection, because the
docker image does not know the final domain of the Testrail instance,
which is required to configure the virtualhost. The Rewrite module used
in .htaccess allows to ignore the current domain.